♻️ Refactor : logic separation
